modified gamePage
