--- connect-4/components/board.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function Board() {
-    return (
-        <div>
-            
-        </div>
-    )
-}
