+++ connect-4/components/gamePageComp/board.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Board = () => {
+  return (
+    <div>
+      This is Board
+    </div>
+  )
+}
+
+export default Board;
