+++ connect-4/components/gamePageComp/playersTurn.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const PlayersTurn = () => {
+  return (
+    <div>
+      player1's turn
+    </div>
+  )
+}
+
+export default PlayersTurn;
